Fixed Minor Issue

+++ Dev/VideoChatClient/Socket/Socket.cpp
@@ -50,7 +50,7 @@ Socket& operator>>(Socket& socket, Buffer& buffer)
 Socket& operator>>(Socket& socket, PacketHeaders& headers)
 {
 	int32_t bytes_read = recv(socket.m_socket.get(), reinterpret_cast<char*>(&headers),
-	                          static_cast<int>(sizeof(PacketHeaders)), 0);
+	                          sizeof(PacketHeaders), 0);
 
 	if (bytes_read < 0)
 		throw VideoChatClientException(VideoChatClientStatus::SOCKET_RECV_FAILED, GetLastError());
